feat: implement patient dashboard and data fetching functionality
